requirements.txt
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.5.3
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.15.1
MarkupSafe==2.1.2
msgpack==1.0.5
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
psutil==5.9.5
pycparser==2.21
python-dotenv==1.0.0
pywin32==306
pyzmq==25.0.2
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
soupsieve==2.4.1
tomli==2.0.1
typing_extensions==4.6.1
urllib3==2.0.2
Werkzeug==2.3.4
zope.event==4.6
zope.interface==6.0
